Return JSONErrorResponse is unsafe to use within Transaction block (#1955)
Move calls to `jsonapi.JSONErrorResponse()` outside of the `application.Transaction(...)`.

Fixes openshiftio/openshift.io#2687

+++ application/repositories.go
@@ -9,6 +9,6 @@ import (
 
 // SearchRepository encapsulates searching of woritems,users,etc
 type SearchRepository interface {
-	SearchFullText(ctx context.Context, searchStr string, start *int, length *int, spaceID *string) ([]workitem.WorkItem, uint64, error)
-	Filter(ctx context.Context, filterStr string, parentExists *bool, start *int, length *int) ([]workitem.WorkItem, uint64, link.AncestorList, link.WorkItemLinkList, error)
+	SearchFullText(ctx context.Context, searchStr string, start *int, length *int, spaceID *string) ([]workitem.WorkItem, int, error)
+	Filter(ctx context.Context, filterStr string, parentExists *bool, start *int, length *int) ([]workitem.WorkItem, int, link.AncestorList, link.WorkItemLinkList, error)
 }
